bug fixes
According to the readme.md:
'# By default, dir to save the results would be ~/TRACE_results. To change it, please specify:
--save_path /path/to/save/folder
'
but it doesn't work even I specify the save_path, it seems the variable used to be the frame path of video input.

BTW, the output path is /home/sunyu/TRACE_results/ now, even I specify it.

To fix it, add one more input to 'save_results'